Perform reverse DNS lookup in NetMQBeacon.HostName getter, avoiding blocking Configure.
This commit also fixes a potential deadlock when the actor had bindTo == null and wouldn't return a response, and the Configure method would block on ReceiveFrameString indefinitely.

The capitalisation of HostName has been corrected to match the BCL. As we're throwing away obsolete code in this repo at the moment, I didn't add an obsolete facade with the old name.