Make NetMQPoller implement ISynchronizeInvoke.
This interface is used in several APIs where synchronisation is offered, especially those which have been around for a long time in the BCL such as WinForms, WebForms, System.Timers.Timer and System.Diagnostics.Process.

The implementation utilises the TaskScheduler, and as such is not available in the .NET 3.5 build.